chore(xlsx): move buildSeriesFromSource + mergeSeries into chart/series
Move the clone-side `buildSeriesFromSource` (per-series merge) and
`mergeSeries` (per-field merge) from chart-clone.ts into
chart/series.ts now that the per-field resolvers (resolveStroke /
resolveMarker / resolveSmooth / etc.) plus the shared `applyOverride`
helper already live there. The two helpers reference the public
`CloneChartSeriesOverride` type; chart/series.ts imports it via
`import type` to avoid a runtime cycle with chart-clone.ts.

Pure relocation — JSDoc preserved verbatim, no behavior change. All
7362 tests still pass; typecheck clean.
